chore(textfield): extract feedback container types to new file
DSY-2081
